Make credentials configurable
By default laravel-multiauth checks that `admins.active = 1`.

This change allows you to configure the credentials that are passed
along with the username and password by adding a `credentials` setting
to `config/multiauth.php`.

With this you can add additional columns to check, change the type of
the `active` value, or remove the `active` check altogether.